+++ neoforge/src/main/java/de/cech12/usefulhats/platform/NeoForgeRegistryHelper.java
@@ -1,5 +1,7 @@
 package de.cech12.usefulhats.platform;
 
+import de.cech12.usefulhats.Constants;
+import de.cech12.usefulhats.compat.CuriosCompat;
 import de.cech12.usefulhats.init.ModItems;
 import de.cech12.usefulhats.item.AbstractHatItem;
 import de.cech12.usefulhats.platform.services.IRegistryHelper;
@@ -45,6 +47,9 @@ public class NeoForgeRegistryHelper implements IRegistryHelper {
         //if (Services.PLATFORM.isModLoaded(Constants.ACCESSORIES_MOD_ID)) {
         //    AccessoriesCompat.addEquippedHatsToList(entity, stacks);
         //}
+        if (Services.PLATFORM.isModLoaded(Constants.CURIOS_MOD_ID)) {
+            CuriosCompat.addEquippedHatsToList(entity, stacks);
+        }
         return stacks;
     }